requirements.txt
# video-processing-microservice/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Video processing
opencv-python==4.8.0.76
numpy==1.24.0

# Google Cloud AI
google-cloud-aiplatform==1.34.0

# Vector database
pinecone-client==2.2.4

# HTTP client
httpx==0.25.0

# Image processing
Pillow==10.0.0

# Settings
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0